qwen-poc/requirements.txt
fastapi
uvicorn
dashscope
pydantic
requests
python-dotenv
openai
fal-client
